+++ crates/core/src/addon.rs
@@ -5,6 +5,12 @@ use std::cmp::Ordering;
 use std::collections::HashMap;
 use std::path::PathBuf;
 
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum AddonVersionKey {
+    Local,
+    Remote,
+}
+
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct RemotePackage {
     pub version: String,
@@ -103,6 +109,12 @@ pub struct Addon {
     #[cfg(feature = "gui")]
     pub details_btn_state: iced_native::button::State,
     #[cfg(feature = "gui")]
+    pub remote_btn_state: iced_native::button::State,
+    #[cfg(feature = "gui")]
+    pub local_btn_state: iced_native::button::State,
+    #[cfg(feature = "gui")]
+    pub full_changelog_btn_state: iced_native::button::State,
+    #[cfg(feature = "gui")]
     pub update_btn_state: iced_native::button::State,
     #[cfg(feature = "gui")]
     pub force_btn_state: iced_native::button::State,
@@ -154,6 +166,12 @@ impl Addon {
             #[cfg(feature = "gui")]
             details_btn_state: Default::default(),
             #[cfg(feature = "gui")]
+            remote_btn_state: Default::default(),
+            #[cfg(feature = "gui")]
+            local_btn_state: Default::default(),
+            #[cfg(feature = "gui")]
+            full_changelog_btn_state: Default::default(),
+            #[cfg(feature = "gui")]
             update_btn_state: Default::default(),
             #[cfg(feature = "gui")]
             force_btn_state: Default::default(),
@@ -191,6 +209,12 @@ impl Addon {
             game_version: None,
             #[cfg(feature = "gui")]
             details_btn_state: Default::default(),
+            #[cfg(feature = "gui")]
+            remote_btn_state: Default::default(),
+            #[cfg(feature = "gui")]
+            local_btn_state: Default::default(),
+            #[cfg(feature = "gui")]
+            full_changelog_btn_state: Default::default(),
             #[cfg(feature = "gui")]
             update_btn_state: Default::default(),
             #[cfg(feature = "gui")]